Add files via upload
Heart disease is the number one cause of death worldwide, so if you're looking to use data science for good you've come to the right place. To learn how to prevent heart disease we must first learn to reliably detect it.
Our dataset is from a study of heart disease that has been open to the public for many years. The study collects various measurements on patient health and cardiovascular statistics, and of course makes patient identities anonymous.
 
About:
Preventing heart disease is important. Good data-driven systems for predicting heart disease can improve the entire research and prevention process, making sure that more people can live healthy lives.
In the United States, the Centers for Disease Control and Prevention is a good resource for information about heart disease. According to their website:
•	About 610,000 people die of heart disease in the United States every year–that’s 1 in every 4 deaths.
•	Heart disease is the leading cause of death for both men and women. More than half of the deaths due to heart disease in 2009 were in men.
•	Coronary heart disease (CHD) is the most common type of heart disease, killing over 370,000 people annually.
•	Every year about 735,000 Americans have a heart attack. Of these, 525,000 are a first heart attack and 210,000 happen in people who have already had a heart attack.
•	Heart disease is the leading cause of death for people of most ethnicities in the United States, including African Americans, Hispanics, and whites. For American Indians or Alaska Natives and Asians or Pacific Islanders, heart disease is second only to cancer.
 
Problem description
Your goal is to predict the binary class heart_disease_present, which represents whether or not a patient has heart disease:
•	0 represents no heart disease present
•	1 represents heart disease present

 